+++ src/ecc/ECC.jl
@@ -1,190 +1,48 @@
 module ECC
 
+using LinearAlgebra
 using QuantumClifford
-using QuantumClifford: AbstractOperation
+using QuantumClifford: AbstractOperation, AbstractStabilizer
 import QuantumClifford: Stabilizer, MixedDestabilizer
+using DocStringExtensions
+using Combinatorics: combinations
 
 abstract type AbstractECC end
 
-"""The encoding circuit of a given code."""
-function encoding_circuit end
+export Shor9, Steane7, Cleve8, Perfect5, Bitflip3,
+    parity_checks, naive_syndrome_circuit, shor_syndrome_circuit, naive_encoding_circuit,
+    code_n, code_s, code_k, rate, distance,
+    isdegenerate, faults_matrix
 
 """Parity check tableau of a code."""
 function parity_checks end
 
+parity_checks(s::Stabilizer) = s
 Stabilizer(c::AbstractECC) = parity_checks(c)
-MixedDestabilizer(c::AbstractECC) = MixedDestabilizer(Stabilizer(c))
+MixedDestabilizer(c::AbstractECC; kwarg...) = MixedDestabilizer(Stabilizer(c); kwarg...)
 
 """The number of physical qubits in a code."""
 function code_n end
 
+code_n(c::AbstractECC) = code_n(parity_checks(c))
+
+code_n(s::Stabilizer) = nqubits(s)
+
 """The number of stabilizer checks in a code."""
-function code_s(c::AbstractECC)
-    s = length(parity_checks(c))
-    return s
-end
+function code_s end
+
+code_s(c::AbstractECC) = code_s(parity_checks(c))
+code_s(s::Stabilizer) = length(s)
 
 """The number of logical qubits in a code."""
-function code_k(c::AbstractECC)
-    k = code_n(c) - code_s(c)
-    return k
-end
+code_k(c) = code_n(c) - code_s(c)
 
 """The rate of a code."""
-function rate(c::AbstractECC)
-    rate = code_k(c)//code_s(c)
+function rate(c)
+    rate = code_k(c)//code_n(c)
     return rate
 end
 
-"""Number of physical qubits for a given parity check tableau"""
-function code_n(parity_check_tableau)
-    return size(parity_check_tableau)[2]
-end
-
-"""Generate the non-fault-tolerant stabilizer measurement cicuit for a given code instance or parity check tableau.
-
-Use the `ancillary_index` and `bit_index` arguments to offset where the corresponding part the circuit starts.
-
-Returns the circuit, the number of ancillary qubits that were added, and a list of bit indices that will store the measurement results.
-
-See also: [`shor_syndrome_circuit`](@ref)
-"""
-function naive_syndrome_circuit end
-
-function naive_syndrome_circuit(code_type::AbstractECC, ancillary_index=1, bit_index=1)
-    naive_syndrome_circuit(parity_checks(code_type), ancillary_index, bit_index)
-end
-
-"""Naive Pauli measurement circuit using a single ancillary qubit.
-
-Not a fault-tolerant circuit, but useful for testing purposes.
-
-Measures the corresponding `PauliOperator` by using conditional gates into an ancillary
-qubit at index `nqubits(p)+ancillary_index` and stores the measurement result
-into classical bit `bit_index`.
-
-See also: [`naive_syndrome_circuit`](@ref), [`shor_ancillary_paulimeasurement`](@ref)"""
-function naive_ancillary_paulimeasurement(p::PauliOperator, ancillary_index=1, bit_index=1)
-    circuit = AbstractOperation[]
-    numQubits = nqubits(p)
-    for qubit in 1:numQubits
-        if p[qubit] == (1,0)
-            push!(circuit, sXCX(qubit, numQubits + ancillary_index))
-        elseif p[qubit] == (0,1)
-            push!(circuit, sCNOT(qubit, numQubits + ancillary_index))
-        elseif p[qubit] == (1,1)
-            push!(circuit, sYCX(qubit, numQubits + ancillary_index))
-        end
-    end
-    mz = sMRZ(numQubits + ancillary_index, bit_index)
-    push!(circuit, mz)
-
-    return circuit
-end
-
-function naive_syndrome_circuit(parity_check_tableau, ancillary_index=1, bit_index=1)
-    naive_sc = AbstractOperation[]
-    ancillaries = 0
-    bits = 0
-    for check in parity_check_tableau
-        append!(naive_sc,naive_ancillary_paulimeasurement(check, ancillary_index+ancillaries, bit_index+bits))
-        ancillaries +=1
-        bits +=1
-    end
-
-    return naive_sc, ancillaries, bit_index:bit_index+bits-1
-end
-
-"""Generate the Shor fault-tolerant stabilizer measurement cicuit for a given code instance or parity check tableau.
-
-Use the `ancillary_index` and `bit_index` arguments to offset where the corresponding part the circuit starts.
-Ancillary qubits
-
-Returns:
-  - The cat state preparation circuit.
-  - The Shor syndrome measurement circuit.
-  - The number of ancillary qubits that were added.
-  - The list of bit indices that store the final measurement results.
-
-See also: [`naive_syndrome_circuit`](@ref)
-"""
-function shor_syndrome_circuit end
-
-function shor_syndrome_circuit(code_type::AbstractECC, ancillary_index=1, bit_index=1)
-    shor_syndrome_circuit(parity_checks(code_type), ancillary_index, bit_index)
-end
-
-"""Shor's Pauli measurement circuit using a multiple entangled ancillary qubits.
-
-A fault-tolerant circuit.
-
-Measures the corresponding PauliOperator by using conditional gates into multiple ancillary
-entangled qubits starting at index `nqubits(p)+ancillary_index`
-and stores the measurement result into classical bits starting at `bit_index`.
-The final measurement result is the XOR of all the bits.
-
-Returns:
-  - The cat state preparation circuit.
-  - The Shor syndrome measurement circuit.
-  - One more than the index of the last added ancillary qubit.
-  - One more than the index of the last added classical bit.
-
-See also: [`naive_syndrome_circuit`](@ref), [`naive_ancillary_paulimeasurement`](@ref)"""
-function shor_ancillary_paulimeasurement(p::PauliOperator, ancillary_index=1, bit_index=1)
-    init_ancil_index = ancillary_index
-    circuit = AbstractOperation[]
-    measurements = AbstractOperation[]
-    numQubits = nqubits(p)
-    for qubit in 1:numQubits
-        if p[qubit] == (1,0)
-            push!(circuit, sXCZ(qubit, numQubits + ancillary_index))
-        elseif p[qubit] == (0,1)
-            push!(circuit, sZCZ(qubit, numQubits + ancillary_index))
-        elseif p[qubit] == (1,1)
-            push!(circuit, sYCZ(qubit, numQubits + ancillary_index))
-        end
-        if p[qubit] != (0,0)
-            push!(measurements, sMRX(numQubits + ancillary_index, bit_index))
-            ancillary_index +=1
-            bit_index +=1
-        end
-    end
-    circuit = vcat(circuit,measurements)
-
-    cat_state_circuit = AbstractOperation[]
-    push!(cat_state_circuit, sHadamard(numQubits + init_ancil_index))
-    for i in (init_ancil_index+1):(ancillary_index -1)
-        push!(cat_state_circuit, sCNOT(numQubits + (i - 1), numQubits + i))
-    end
-
-    return cat_state_circuit, circuit, ancillary_index, bit_index
-end
-
-function shor_syndrome_circuit(parity_check_tableau, ancillary_index=1, bit_index=1)
-    shor_sc = AbstractOperation[]
-    xor_indices = []
-    cat_circuit = AbstractOperation[]
-    initial_ancillary_index = ancillary_index
-
-    for check in parity_check_tableau
-        cat_circ, circ, new_ancillary_index, new_bit_index = shor_ancillary_paulimeasurement(check, ancillary_index, bit_index)
-        push!(xor_indices, collect(bit_index:new_bit_index-1))
-
-        append!(shor_sc, circ)
-        append!(cat_circuit, cat_circ)
-
-        ancillary_index = new_ancillary_index
-        bit_index = new_bit_index
-    end
-
-    final_bits = 0
-    for indices in xor_indices
-        push!(shor_sc, QuantumClifford.ClassicalXOR(indices, bit_index+final_bits))
-        final_bits += 1
-    end
-
-    return cat_circuit, shor_sc, ancillary_index-initial_ancillary_index, bit_index:bit_index+final_bits-1
-end
 
 """The distance of a code."""
 function distance end
@@ -196,13 +54,13 @@ function parity_matrix(c::AbstractECC)
 end
 
 """Logical X operations of a code."""
-function logx_ops(c::AbstractECC)
+function logx_ops(c)
     md = MixedDestabilizer(parity_checks(c))
     logicalxview(md)
 end
 
 """Logical Z operations of a code."""
-function logz_ops(c::AbstractECC)
+function logz_ops(c)
     md = MixedDestabilizer(parity_checks(c))
     logicalzview(md)
 end
@@ -392,10 +250,50 @@ function faults_matrix(c::AbstractECC)
     return faults_matrix(parity_checks(c))
 end
 
-# TODO implement isdegenerate
+"""
+$TYPEDSIGNATURES
 
-include("./bitflipcode.jl")
-include("./shorcode.jl")
-include("./steanecode.jl")
+Check if the code is degenerate with respect to a given set of error or with respect to all
+"up to d physical-qubit" errors (defaulting to d=1).
+
+```jldoctest
+julia> using QuantumClifford.ECC
+
+julia> isdegenerate(Shor9(), [single_z(9,1), single_z(9,2)])
+true
+
+julia> isdegenerate(Shor9(), [single_z(9,1), single_x(9,1)])
+false
+
+julia> isdegenerate(Steane7(), 1)
+false
+
+julia> isdegenerate(Steane7(), 2)
+true
+```
+"""
+function isdegenerate end
+
+isdegenerate(c::AbstractECC, args...) = isdegenerate(parity_checks(c), args...)
+isdegenerate(c::AbstractStabilizer, args...) = isdegenerate(stabilizerview(c), args...)
+
+function isdegenerate(H::Stabilizer, errors) # Described in https://quantumcomputing.stackexchange.com/questions/27279
+    syndromes = comm.((H,), errors) # TODO This can be optimized by having something that always returns bitvectors
+    return length(Set(syndromes)) != length(errors)
+end
+
+function isdegenerate(H::Stabilizer, d::Int=1)
+    n = nqubits(H)
+    errors = [begin p=zero(PauliOperator,n); for i in bits; p[i]=op; end; p end for bits in combinations(1:n, d) for op in ((true,false), (false,true))]
+    return isdegenerate(H, errors)
+end
+
+include("circuits.jl")
+
+include("codes/bitflipcode.jl")
+include("codes/fivequbit.jl")
+include("codes/steanecode.jl")
+include("codes/shorcode.jl")
+include("codes/clevecode.jl")
 
 end #module